Update spell-selection.component.ts change how special powers are read in when passed in via selectionType.  Add propertyIndex Update character-magic-subtheme.component.html change passing in special powers to account for new values Update constants.ts SpecialPower class now has a requirement (a new enum) and number of selections a number.
Reasoning:  PP, AP and other build specific stuff that was seperate from the special power but apart of the build are now all being rolled into an array of special powers.  The special powers represent the variable chooices a character can make.  At least a character won't have any choices to make and will merely need to select and option.  At most a character will need to make a 1 of 3 and another 1 of two choices.  Since all choices are stored in the special powers array we need to filter only valid choices see special powers Requirements and then make sure to load only those filtered changes when we are in the spell selection class see propertyIndex.

TODO: adjustment will need to be made to include all information in a build what is effectively 1 spell selection.  This can be accomplished by adding to things like On Hit and Spell Effect, the text related to Power point ability and rare power/ AP power.  Other bonus can be added without a table if one isn't needed, like a warlock's +1 to saving throw.  Once the selection is made all powers of the build will be added to the character

TODO: fix all broken spell selection tests as a result of this change